requirements.txt
aiogram>=3.17.0
tortoise-orm>=0.23.0
asyncpg>=0.30.0
pyyaml>=6.0.2
python-dotenv>=1.0.1
aerich>=0.7.2